+++ zaino-state/src/error.rs
@@ -2,7 +2,7 @@
 
 use crate::Hash;
 
-use std::any::type_name;
+use std::{any::type_name, fmt::Display};
 
 use zaino_fetch::jsonrpsee::connector::RpcRequestError;
 
@@ -463,3 +463,51 @@ pub enum FinalisedStateError {
 #[derive(Debug, Clone, thiserror::Error)]
 #[error("Unexpected status error: {0:?}")]
 pub struct StatusError(pub crate::status::StatusType);
+
+#[derive(Debug, thiserror::Error)]
+#[error("{kind}: {message}")]
+/// The set of errors that can occur during the public API calls
+/// of a NodeBackedChainIndex
+pub struct ChainIndexError {
+    pub(crate) kind: ChainIndexErrorKind,
+    pub(crate) message: String,
+    pub(crate) source: Option<Box<dyn std::error::Error + Send + Sync + 'static>>,
+}
+
+#[derive(Debug)]
+/// The high-level kinds of thing that can fail
+pub enum ChainIndexErrorKind {
+    /// Zaino is in some way nonfunctional
+    InternalServerError,
+    /// The given snapshot contains invalid data.
+    InvalidSnapshot,
+}
+
+impl Display for ChainIndexErrorKind {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_str(match self {
+            ChainIndexErrorKind::InternalServerError => "internal server error",
+            ChainIndexErrorKind::InvalidSnapshot => "invalid snapshot",
+        })
+    }
+}
+
+impl ChainIndexError {
+    pub(crate) fn backing_validator(value: impl std::error::Error + Send + Sync + 'static) -> Self {
+        Self {
+            kind: ChainIndexErrorKind::InternalServerError,
+            message: "InternalServerError: error receiving data from backing node".to_string(),
+            source: Some(Box::new(value)),
+        }
+    }
+
+    pub(crate) fn database_hole(missing_block: impl Display) -> Self {
+        Self {
+            kind: ChainIndexErrorKind::InternalServerError,
+            message: format!(
+                "InternalServerError: hole in validator database, missing block {missing_block}"
+            ),
+            source: None,
+        }
+    }
+}
